backend/requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Geospatial libraries
shapely>=2.0.0
geopandas>=0.14.0
pyproj>=3.6.0
geojson>=3.0.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# AI/ML - Segment Anything Model (SAM)
segment-anything>=1.0
torch>=2.0.0
torchvision>=0.15.0
scikit-image>=0.21.0
scipy>=1.11.0

# AI Services
google-generativeai>=0.3.0

# HTTP requests for satellite tiles
requests>=2.31.0
aiohttp>=3.9.0

# PDF generation
reportlab>=4.0.0
matplotlib>=3.8.0

# Utilities
python-dotenv>=1.0.0
